@@ -2,26 +2,25 @@
   "name": "mlp-client",
   "version": "1.4.0",
   "main": "./src/App.js",
-  "type": "module",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "leaflet-kml": "^1.0.1",
     "file-saver": "^2.0.5",
-    "mathjs": "^9.4.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.2",
-    "react-flatpickr": "^3.10.7",
+    "mathjs": "^11.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-flatpickr": "^3.10.13",
     "react-helmet": "^6.1.0",
     "react-idle-timer": "^4.5.6",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.1",
     "utif": "^3.1.0"
   },
   "devDependencies": {
-    "jest": "^26.6.0",
-    "web-vitals": "^0.2.4"
+    "jest": "^29.4.3",
+    "web-vitals": "^3.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
